# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a doc for compass_mini_api.",
        "title": "compass_mini_api",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth/get_data_token": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Data Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get Data Token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetDataTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "description": "Login user, get your Geolocation from https://ip.nf/me.json",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "type": "string",
                        "example": "Android",
                        "description": "entity",
                        "name": "entity",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AuthLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AuthLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Logout user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AuthLogoutResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/splash": {
            "post": {
                "description": "Splash screen",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Splash screen",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AuthSplashRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AuthSplashResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AuthLoginRequest": {
            "type": "object",
            "required": [
                "geolocation",
                "mobilephone",
                "password"
            ],
            "properties": {
                "geolocation": {
                    "$ref": "#/definitions/dto.Geolocation"
                },
                "mobilephone": {
                    "type": "string",
                    "example": "+6281234567890"
                },
                "password": {
                    "type": "string",
                    "example": "Test12345@"
                }
            }
        },
        "dto.AuthLoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "your access token"
                },
                "refresh_token": {
                    "type": "string",
                    "example": "your refresh token"
                }
            }
        },
        "dto.AuthLogoutResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.AuthSplashRequest": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "your access token"
                },
                "refresh_token": {
                    "type": "string",
                    "example": "your refresh token"
                }
            }
        },
        "dto.AuthSplashResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "your access token"
                },
                "refresh_token": {
                    "type": "string",
                    "example": "your refresh token"
                }
            }
        },
...
        "dto.Geolocation": {
            "type": "object",
            "properties": {
                "asn": {
                    "type": "string",
                    "example": "AS23700 Linknet-Fastnet ASN"
                },
                "city": {
                    "type": "string",
                    "example": "Jakarta"
                },
                "country": {
                    "type": "string",
                    "example": "Indonesia"
                },
                "country_code": {
                    "type": "string",
                    "example": "ID"
                },
                "hostname": {
                    "type": "string",
                    "example": "fm-dyn-111-94-121-97.fast.net.id."
                },
                "ip": {
                    "type": "string",
                    "example": "111.94.121.97"
                },
                "latitude": {
                    "type": "number",
                    "example": -6.1743998527526855
                },
                "longitude": {
                    "type": "number",
                    "example": 106.82939910888672
                },
                "netmask": {
                    "type": "integer",
                    "example": 16
                },
                "post_code": {
                    "type": "string",
                    "example": "15710"
                }
            }
        },
...
        "dto.GetDataTokenResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "identityid": {
                    "type": "integer"
                },
                "isactive": {
                    "type": "boolean"
                },
                "islocked": {
                    "type": "boolean"
                },
                "mobilephone": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "roleid": {
                    "type": "integer"
                }
            }
        },
...
        "response.Meta": {
            "type": "object",
            "properties": {
                "error_id": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean",
                    "default": true
                }
            }
        },
        "response.errorResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "description": {},
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                }
            }
        }
    }
}